added more variables
added a listener which will activate the result function once the button is clicked. made a new variable to store storyText in order to make new random stories and not just be limited to one use. and added variables that will take a random string from a specific array and replace it with the place holder in newStory